chore: Refactored JPA classes into a JPA package to keep it all in one place
