+++ Backend/src/main/java/com/ITProject/RCPT/Controllers/LookupControllers/StipendController.java
@@ -1,7 +1,7 @@
 package com.ITProject.RCPT.Controllers.LookupControllers;
 
-import com.ITProject.RCPT.Entities.Stipend;
-import com.ITProject.RCPT.Services.StipendService;
+import com.ITProject.RCPT.JPA.Entities.Stipend;
+import com.ITProject.RCPT.JPA.Services.StipendService;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
