+++ Backend/src/main/java/com/ITProject/RCPT/JPA/Entities/DepartmentAndFaculty.java
@@ -1,4 +1,4 @@
-package com.ITProject.RCPT.Entities;
+package com.ITProject.RCPT.JPA.Entities;
 
 import jakarta.persistence.*;
