+++ Backend/src/main/java/com/ITProject/RCPT/JPA/Repositories/DepartmentAndFacultyRepository.java
@@ -1,6 +1,6 @@
-package com.ITProject.RCPT.Repositories;
+package com.ITProject.RCPT.JPA.Repositories;
 
-import com.ITProject.RCPT.Entities.DepartmentAndFaculty;
+import com.ITProject.RCPT.JPA.Entities.DepartmentAndFaculty;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
