+++ Backend/src/main/java/com/ITProject/RCPT/JPA/Repositories/RegionRepository.java
@@ -1,6 +1,6 @@
-package com.ITProject.RCPT.Repositories;
+package com.ITProject.RCPT.JPA.Repositories;
 
-import com.ITProject.RCPT.Entities.Region;
+import com.ITProject.RCPT.JPA.Entities.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
