+++ Backend/src/main/java/com/ITProject/RCPT/JPA/Services/StaffBenefitsService.java
@@ -1,7 +1,7 @@
-package com.ITProject.RCPT.Services;
+package com.ITProject.RCPT.JPA.Services;
 
-import com.ITProject.RCPT.Entities.StaffBenefits;
-import com.ITProject.RCPT.Repositories.StaffBenefitsRepository;
+import com.ITProject.RCPT.JPA.Entities.StaffBenefits;
+import com.ITProject.RCPT.JPA.Repositories.StaffBenefitsRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
